% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plot.R
\name{quant_plot}
\alias{quant_plot}
\title{A quantile plot function}
\usage{
quant_plot(d, ylabel = "Quantity", quantile = seq(0, 1, by = 0.25))
}
\arguments{
\item{ylabel}{Label for the quantity being plotted; defaults to "Quantity"}

\item{quantile}{Vector of numeric vector of probabilities with values in [0,1]
representing the quantile boundaries. This defaults to quartiles seq(0,1,by=.25).}

\item{data}{An atomic vector of numbers}
}
\description{
This function takes an atomic vector of numbers and
creates a quantile plot of the data. This function requires
GGally, ggplot2, reshape2 (probably)
}
\examples{
quant_plot(data=df$height,ylabel="Height")
}
\keyword{plot,}
\keyword{quant_plot}
\keyword{quantile}

