% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quant_plot.R
\name{quant_plot}
\alias{quant_plot}
\title{A quantile plot function}
\usage{
quant_plot(d, ylabel = "Quantity", breaks = seq(0, 1, 0.25))
}
\arguments{
\item{ylabel}{Label for the quantity being plotted; defaults to "Quantity"}

\item{breaks}{Vector of numeric vector of probabilities with values in [0,1] representing the quantile boundaries. CURRENTLY IGNORED}

\item{data}{An atomic vector of numbers}
}
\description{
This function takes an atomic vector of numbers and
creates a quantile plot of the data.
}
\examples{
quant_plot(data=x,ylabel="Height")
}
\keyword{quant_plot}

